requirements.txt
asgiref==3.6.0
certifi==2022.12.7
chardet==3.0.4
Django==4.2
django-ckeditor==6.5.1
django-cors-headers==3.14.0
django-cuser==2017.3.16
django-js-asset==2.0.0
django-modeltranslation==0.18.9
django-storages==1.13.2
googletrans==3.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
Pillow==9.5.0
rfc3986==1.5.0
sniffio==1.3.0
sqlparse==0.4.4
typing_extensions==4.5.0
